requirements.txt
aiofiles==22.1.0
anyio==3.6.1
certifi==2022.6.15.2
click==8.1.3
fastapi==0.83.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
numpy==1.23.3
opencv-python==4.6.0.66
pydantic==1.10.2
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
web3==5.30.0
Pillow==9.2.0
python-dotenv==0.21.0